Remove unused shield-study-addon-utils
This library is commonly used in Shield studies. This add-on is not a
Shield study (since it is intended to be released via Balrog) and does
not use the study utils for anything.

Removing this library reduces the size of the built XPI by over 80%, and
decreases the time it takes to build the add-on.